+++ src/processor.cpp
@@ -1,7 +1,8 @@
 #include "processor.h"
 #include <string>
 
-// TODO: Return the aggregate CPU utilization
+// Return the aggregate CPU utilization
+// Calculated since the last time this function was called
 float Processor::Utilization() 
 { 
     float utilization;
@@ -12,14 +13,19 @@ float Processor::Utilization()
 
     current_total = LinuxParser::Jiffies();
     current_active = LinuxParser::ActiveJiffies();
+
+    // difference in total jiffies
     d_total = current_total - prev_total_;
+
+    // difference in active jiffies
     d_active = current_active - prev_active_;
 
-    // update private members
+    // update
     prev_total_ = current_total;
     prev_active_ = current_active;
-
-    utilization = (float)(d_active) / d_total;
+    
+    utilization = (float)(d_active) / (float)d_total;
+    
     return utilization;
 
 }
